Moved benchmark files, and cleaned :D
