+++ Benchmark/Engine.Benchmark/Program.cs
@@ -7,7 +7,9 @@ public static class Program
 {
     public static async Task Main()
     {
-        //BenchmarkRunner.Run<PolilineBufferBenchmark>();
         BenchmarkRunner.Run<StationsAroundPolyline>();
+        BenchmarkRunner.Run<Polyline6Decode>();
+        BenchmarkRunner.Run<Polyline6DecodeParallel>();
+        BenchmarkRunner.Run<OsrmRouterBenchmark>();
     }
 }
